src/requirements.txt
# protobuf==3.5.2.post1
#protobuf>=3.5
protobuf>=3.12
absl-py>=0.9
